feat(events): implement event handlers for restaurant creation, deletion, and update feat(migrations): add documentation for EF Core migrations refactor: update messaging references and improve RabbitMQ integration
