+++ src/Restaurant/RestoRate.Restaurant.Domain/RestaurantAggregate/Events/RestaurantCreatedEvent.cs
@@ -1,10 +1,8 @@
 using Ardalis.SharedKernel;
 
-using Mediator;
-
 namespace RestoRate.Restaurant.Domain.RestaurantAggregate.Events;
 
-public sealed class RestaurantCreatedEvent(int restaurantId, string name) : DomainEventBase, INotification
+public sealed class RestaurantCreatedEvent(int restaurantId, string name) : DomainEventBase
 {
     public int RestaurantId { get; init; } = restaurantId;
     public string Name { get; init; } = name;
