+++ src/Restaurant/RestoRate.Restaurant.Domain/RestaurantAggregate/Events/RestaurantDeletedEvent.cs
@@ -1,9 +1,8 @@
 using Ardalis.SharedKernel;
-using Mediator;
 
 namespace RestoRate.Restaurant.Domain.RestaurantAggregate.Events;
 
-public sealed class RestaurantDeletedEvent(int restaurantId) : DomainEventBase, INotification
+public sealed class RestaurantDeletedEvent(int restaurantId) : DomainEventBase
 {
     public int RestaurantId { get; init; } = restaurantId;
 }
